requirements.txt
Django==4.2.5
django-environ==0.10.0
django-filter==23.1
django-mysql==3.12.0
django-sendfile2==0.7.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
markdown==3.5.2
gunicorn==21.2.0
drf-yasg==1.21.5
mysqlclient==2.2.0
PyJWT==2.8.0
numpy==1.26.0
python-dateutil==2.8.2
xhtml2pdf==0.2.9
reportlab==4.0.4
Pillow==10.1.0
matplotlib==3.8.0
pandas==2.1.1
openpyxl==3.1.2
plotly==5.18.0
kaleido==0.2.1
qrcode[pil]==7.4.2
cryptography==41.0.4
Babel==2.13.1